Move DefaultConfigProperties to autoconfigure-spi
