+++ sdk-extensions/autoconfigure/src/test/java/io/opentelemetry/sdk/autoconfigure/ConfigPropertiesTest.java
@@ -13,6 +13,7 @@ import static org.assertj.core.api.Assertions.entry;
 
 import io.opentelemetry.sdk.autoconfigure.spi.ConfigProperties;
 import io.opentelemetry.sdk.autoconfigure.spi.ConfigurationException;
+import io.opentelemetry.sdk.autoconfigure.spi.internal.DefaultConfigProperties;
 import java.time.Duration;
 import java.util.Arrays;
 import java.util.Collections;
@@ -236,7 +237,7 @@ class ConfigPropertiesTest {
     expectedMap.put("bear", "growl");
 
     Map<String, String> map = makeTestProps();
-    ConfigProperties properties = DefaultConfigProperties.get(map);
+    ConfigProperties properties = DefaultConfigProperties.create(map);
     assertThat(properties.getBoolean("test.boolean", false)).isTrue();
     assertThat(properties.getString("test.string", "nah")).isEqualTo("str");
     assertThat(properties.getDouble("test.double", 65.535)).isEqualTo(5.4);
@@ -250,7 +251,7 @@ class ConfigPropertiesTest {
 
   @Test
   void defaultMethodsFallBack() {
-    ConfigProperties properties = DefaultConfigProperties.get(emptyMap());
+    ConfigProperties properties = DefaultConfigProperties.create(emptyMap());
     assertThat(properties.getBoolean("foo", true)).isTrue();
     assertThat(properties.getString("foo", "bar")).isEqualTo("bar");
     assertThat(properties.getDouble("foo", 65.535)).isEqualTo(65.535);
@@ -261,7 +262,7 @@ class ConfigPropertiesTest {
 
   @Test
   void defaultCollectionTypes() {
-    ConfigProperties properties = DefaultConfigProperties.get(emptyMap());
+    ConfigProperties properties = DefaultConfigProperties.create(emptyMap());
     assertThat(properties.getList("foo", Arrays.asList("1", "2", "3")))
         .containsExactly("1", "2", "3");
     assertThat(properties.getList("foo")).isEmpty();
